Restructuring codebase slightly, scripts folder, planner folders, put stuff out of utils and either in src directly or in subfolders
