+++ src/planners/straight_line_planner.py
@@ -1,4 +1,4 @@
-from src.utils.planner import Planner
+from src.planners.planner import Planner
 import math
 import numpy as np
